requirements.txt
defusedxml==0.7.1
Django==4.2.27
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.2
inflection==0.5.1
psycopg2-binary==2.9.10
six==1.16.0
sqlparse>=0.5.0 # not directly required, pinned by Snyk to avoid a vulnerability
gunicorn==21.2.0
whitenoise==6.6.0
django-redis==5.4.0
redis==5.0.7
